functional Paginator component
